[063] add config var for episode metadata feature
the var is phrased in the negative to work more easily with env vars.
unless env vars are parsed they are strings, so setting the varaible to
any value will be considered "truthy". Therefore it is easiest to look
for a flag that says NOT to set the metadata, rather than look for a
flag that DOES say to set it.